Assert bootstrap store better, move use_store tests where it belongs
